+++ tests/testthat/test_checkDataTable.R
@@ -56,3 +56,14 @@ test_that("list columns", {
   expect_true(testDataTable(x, any.missing = FALSE))
   expect_true(testDataTable(x, all.missing = FALSE))
 })
+
+test_that("nrow for null data tables", {
+  # c.f. https://github.com/Rdatatable/data.table/issues/3149
+  M = matrix(1:3, nrow = 3)
+  M = M[, integer(0)]
+  DT = data.table::as.data.table(M) # null data table
+
+  expect_equal(rownames(DT), as.character(1:3))
+  expect_equal(nrow(DT), 0)
+  expect_true(testDataTable(DT, nrow = 0))
+})