+++ R/makeExpectation.R
@@ -45,24 +45,31 @@ makeExpectation = function(x, res, info, label) {
 #' @template use.namespace
 #' @export
 makeExpectationFunction = function(check.fun, c.fun = NULL, use.namespace = FALSE, env = parent.frame()) {
-  fn.name = if (!is.character(check.fun)) deparse(substitute(check.fun)) else check.fun
+  fun.name = if (!is.character(check.fun)) deparse(substitute(check.fun)) else check.fun
   check.fun = match.fun(check.fun)
-  fargs = formals(args(check.fun))
+  check.args = fun.args = formals(args(check.fun))
+  x.name = names(fun.args[1L])
   x = NULL
 
   new.fun = function() TRUE
-  if (use.namespace) {
-    formals(new.fun) = c(fargs, alist(info = NULL, label = checkmate::vname(x)))
-    tmpl = "{ res = %s(%s); checkmate::makeExpectation(x, res, info, label) }"
-  } else {
-    formals(new.fun) = c(fargs, alist(info = NULL, label = vname(x)))
-    tmpl = "{ res = %s(%s); makeExpectation(x, res, info, label) }"
-  }
+  body = sprintf("if (missing(%s)) stop(sprintf(\"Argument '%%s' is missing\", .var.name))", x.name)
+
   if (is.null(c.fun)) {
-    body(new.fun) = parse(text = sprintf(tmpl, fn.name, paste0(names(fargs), collapse = ", ")))
+    body = paste0(body, sprintf("; res = %s(%s)", fun.name, paste0(names(check.args), collapse = ", ")))
   } else {
-    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(fargs)), collapse = ", ")))
+    body = paste0(body, sprintf("; res = .Call(%s)", paste0(c(c.fun, names(check.args)), collapse = ", ")))
   }
+
+  if (use.namespace) {
+    formals(new.fun) = c(fun.args, alist(info = NULL, label = checkmate::vname(x)))
+    body = paste0(body, "; checkmate::makeExpectation")
+  } else {
+    formals(new.fun) = c(fun.args, alist(info = NULL, label = vname(x)))
+    body = paste0(body, "; makeExpectation")
+  }
+  body = paste0(body, sprintf("(%s, res, info, label)", x.name))
+
+  body(new.fun) = parse(text = paste("{", body, "}"))
   environment(new.fun) = env
   return(new.fun)
 }