+++ R/makeTest.R
@@ -39,15 +39,19 @@ makeTestFunction = function(check.fun, c.fun = NULL, env = parent.frame()) {
   fun.name = if (is.character(check.fun)) check.fun else deparse(substitute(check.fun))
   check.fun = match.fun(check.fun)
   fun.args = formals(args(check.fun))
+  x.name = names(fun.args[1L])
 
   new.fun = function() TRUE
   formals(new.fun) = fun.args
+  body = sprintf("if (missing(%s)) stop(sprintf(\"Argument '%%s' is missing\", .var.name))", x.name)
   tmpl = "{ isTRUE(%s(%s)) }"
   if (is.null(c.fun)) {
-    body(new.fun) = parse(text = sprintf(tmpl, fun.name, paste0(names(fun.args), collapse = ", ")))
+    body = paste0(body, "; isTRUE(", fun.name, "(", paste0(names(fun.args), collapse = ", "), "))")
   } else {
-    body(new.fun) = parse(text = sprintf(tmpl, ".Call", paste0(c(c.fun, names(fun.args)), collapse = ", ")))
+    body = paste0(body, "; isTRUE(.Call(", paste0(c(c.fun, names(fun.args)), collapse = ", "), "))")
   }
+
+  body(new.fun) = parse(text = paste("{", body, "}"))
   environment(new.fun) = env
   return(new.fun)
 }